Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-09.html

<!DOCTYPE html>
<html>
<head>
<script src='resources/multiple-iframe-test.js'></script>
<script>
var tests = [
    ['yes', 'script-src http://*.0.1:8000', 'resources/script.js'],
    ['yes', 'script-src http://*.0.1:8000/', 'resources/script.js'],
    ['yes', 'script-src http://*.0.1:*', 'resources/script.js'],
    ['yes', 'script-src http://*.0.1:*/', 'resources/script.js'],
];
</script>
</head>
<body onload="test()">
<p>Host wildcards should work correctly.</p>
