LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-09-expected.txt

Host wildcards should work correctly.



--------
Frame: '<!--frame1-->'
--------
PASS

--------
Frame: '<!--frame2-->'
--------
PASS

--------
Frame: '<!--frame3-->'
--------
PASS

--------
Frame: '<!--frame4-->'
--------
PASS
